Avoid duplicate uplink sequence numbers when receiving bad MIC or address in confirmed transmits.
If during a confirmed message transmission we receive a downlink frame with either
an address not matching ours, or an incorrect MIC, control is returned to the user
application with an error status, and the upLinkCounter is not incremented. Thus,
the next uplink frame is sent with the same sequence number, and likely discarded
by the receiving server.

The fix involves discarding the invalid downlink and continuing the retransmission
of the uplink frame until either a downlink response/ack is received, or the retry
count is reached.