Prevent AddMacCommand() from overwriting innocent memory.
When adding a MAC command or response, there was no check in place to
ensure the MAC command buffer was large enough to hold all parameters
associated with a command before writing those parameters to the buffer.

A check was made after the fact, but even though a buffer full status
was returned to the calling function, the damage was done.

This change adds a constant for the MAC buffer size and checks the MAC
buffer has sufficient room for the MAC command or response byte, and
its parameters.