Change initial UpLinkCounter values to 0
In LoRaWAN specs (1.0.2), the frame counter is supposed to be reset to 0 on start of APB session. In the code the initial values was 1.
4.3.1.5 Frame counter (FCnt)
 
Lines 21/22 After a JoinReq – JoinAccept message exchange or a reset for a personalized end-device, the frame counters on the end-device and the frame counters on the network server for that end-device are reset to 0.