Increased default timeout for RegionEU868 to 4000 ms
For SF12 packets of 51 bytes, the TxTimeOnAir is ~2800 ms. With a few additional delays (in the Radio and firmware) the default timeout of 3000 is sometimes reached, leading to the invalid conclusion that Tx has timed out, while the packet has been sent. This, in turn, leads to the possibility of a new packet being sent in violation of dutycycle constraints.

A better approach would be to take the calculated TxTimeOnAir plus 1000 ms as timeout, but that would require a bit of rework on both the region code as well as on the Radio code.