Added a flag to make encryption state of message explicit in LoRaMacCryptoSecureMessage
In rare cases, e.g. when a Tx has timed out, the CryptoNvm context will keep an invalid (already incremented) value for fcntUp, leading to the situation that the next message is sent in clear text, and subsequently received as garbage by the unknowing backend system that attempts to decrypt it.

This patch makes the encrypion state explicit in struct sLoRaMacMessageData.