LoRaMacSerializerData buffer overflow fix
A buffer size is computed and used to check the buffer size, but two flaws introduce the possibility of memory corruption:
- the MIC field is not taken into account
- the FOpts length is only taken into account when payloadsize>0, while the FOpts are always added

With this patch the correct number of bytes is calculated.
The calculation looks a bit verbose now, but for clarity it follows the exact order in which the message will be built up. Compiler optimization should combine all constants into one compound value.