Temporarily remove the HAL_PWR_EnterSLEEPMode call when compiling in debug mode.
Due to an yet unknown reason the PWR_MAINREGULATOR_ON constant gets
changed inside the function which makes the assert fail.

When compiling in release mode the code operates as expected.

TODO: Check what causes this issue. First guess is that the stack gets
      corrupted somehow.

This function is only called when using the GPS peripheral.