soft-se/cmac: quick fix for misaligned 32-bit access
AES_CMAC_CTX.X is not aligned to a 32-bit boundary (offset is 91), this
is an issue when using AES with HAVE_UINT_32T support because it
requires 32-bit aligned buffers.

Fix it the same ugly way it was already done for other calls to
aes_encrypt() by copying the buffer to the stack. Proper fix is to set a
struct member alignment requirement to 32-bit by using __attribute__
((aligned (4)) but it's not portable.