Fix incorrect credentials (#76)
It was not assigning the current FNwkSIntKey, SNwkSIntKey, NwkSEncKey, AppSKey or DevAddress. Put in some tests to test correct assignment.  Needs more tests.

But there is an issue, where the app state is not reset between tests, so each test is linked to others.


* correct key assignment

* set all default lorawan parameters to zeros

So that the unittester can verify if its all correct.

* set dev address from network key selector

* test to see if lorawan credentials are set correctly

* make CommissioningParams externally accessible

For Unittesting

* do network keys test in its own file

* Revert accidentally formatting crypto