Compress only non changing values when saving to EEPROM(#115)
The problem is, saving to EEPROM is slow. If we compress the whole nvm, each time we compress, even if one or 2 bytes are different, the diff between the existing image and one to write is very large. It could be more than 500 bytes difference, taking 2-5 seconds to write.

On the other hand, if we don't compress, and update the changed bytes with each TX, it only changes around 30 bytes in 200 ms or so. But that takes up a lot of space. 2.2kb of space per region(US, CN, EU etc). This is greater than the available space of 6 kb.

So this is the hybrid approach: Compress all those bytes that normally don't change and store them. Don't compress those bytes that change with each tx, updating only those few bytes with each tx. 

The result is, it takes only around 500 ms to write/check the 30 or so bytes changed with each tx, while keeping the whole nvm image under 1 kbytes.