Write words instead of individual bytes to EEPROM (#122)
It takes the same time to write 4 bytes(word) as it takes to write 1 byte. So write 4 bytes at a time. This brings write time from ~250 ms down to ~90 ms to update NVM. Major time savings.

put in mocking function to simulate writing word to eeprom

It was not implemented earlier

Update eeprom-board-mock.cpp

create seperate function to write word and byte

Or else, the EEPROM wipe function is unable to wipe.
TODO: use the same function. see #128

Write twice to eeprom because first write does not work for some reason

Need to find out why this happens. It seems to be code related because sometimes it does not happen when code changes.