Add crc to playback params (#157)
This commit adds in a CRC check on the key playback params:  n_playback_positions_saved and current_EEPROM_index. These need to be guarded against corruption, which can happen during EEPROM write. If incorrect, it could get stuck in infinite loop trying to read the eeprom memory. 

The unittest verifies if the allocated space for this CRC and params is correct at runtime.