Optimise gps driver code for low peak currents (#188)
Rewrites to the Ublox GPS driver code to ensure it does go into low power. Verified with a power analyser. Peak currents are 40 mA when both processor and GPS is on. Currently, max duration in this 40 mA state is 50 milliseconds. Other peaks are shorter, less than 10 ms when sending and receiving data from GPS. Otherwise, the peak current is 25 mA.

The code has also been stripped of all unnecessary functions to make it cleaner.