Join ahead of launch (#209)
Simply switch to using ABP for both Things Network and Helium Network. On The Helium network, it can only use OTAA, but we hardcode in the DevAddr, AppEUI, NwkSKey and AppSKey and make it join with ABP activation. It works.

Implements a counter in EEPROM to ensure frame count always increments.

This commit removes all the functionality to save the whole LoRaMAC state, including the complex compression. Only thing that is saved is frame counter, one for each network.

TODO: make frame counter saving more robust.

autoformat