Cycle through list of credentials in each country (#219)
This is one workaround for #218. In Japan, it cycles through device credentials registered on the EU1 cluster and the AU1 Things Network cluster(and the Helium network). I am hoping all gateways in Japan will be on either the EU1 or AU1.

Code also registers devices on US1 for US credentials, and AU1 for Australian credentials. CN, IN, KR, RU, EU remain on EU1 cluster.