Add in multiple device registrations to reduce fcount increments (#244)
In order to reduce the rate of frame count increase for devices,  register multiple devices, and cycle through all of them, incrementing the frame count for each device one by one. This will reduce the rate of frame count increase, so that server and end node frame count does not diverge more than 2^14. See #210 for more info on the problem this commit solves.

* add in structures for helium network credentials

* put in the remaining credentials for helium 2-5

* put in 5 more device registrations

* number the devices

* update eu keys to use the multiple helium devices

* put in credentials for helium 6-10

* put in new key for helium 1

it was not working earlier for some reason

* put in new eu keys

* put in structures

* put in more devices

* updates keys for the other registered devices ttn

* use a common frame count start for each network

* put in structure for us keys

* update keys for us915

* update comment

* put in basic sanity size check

* put device list for US

* make tests compile

They don't pass though

Update lorawan_credentials_ut.cpp

* spread device frame count usage

* alternate helium keys over russia

* use better order over Japan

* update bug discription comment

* update comment