Loop through more unique helium device registrations (#248)
Loop more through more credentials. This allows future changes to the credentials, while still looping through multiple devices. For example, if we in the future register the device on an another network(that is built in the future), we can send those new credentials to the tracker to occupy these slots. We can still loop through the current existing registrations as well.

* loop through more helium credentials

* more looping