Make watchdog catch timer failures (#255)
In the event the wakeup timer for the next event is NOT set, it could be in deep sleep forever, kicking the watchdog timer as designed. The watchdog should instead reset when the next expected event(e.g. transmit) does not happen.

So just remove this watchdog kick timer. Its a weak point.

But as a compromise, the longest sleep delay is 32 seconds. This lenght of time is the longest possible watchdog reset interval. So we schedule transmissions to happen every 31 seconds, down from 60 seconds.


* watchdog reset IMMEDIATELY after wakeup

* reduce tx interval down to 31 seconds

So that the watchdog won't kickin before that