Tx with new gps fix each time (#262)
This commit significantly simplifies the code in `main.c`. Then makes it get a GPS fix with every single transmission. The interval between transmissions are set depending whether a GPS fix was successful or not. If not successful, resume search in 10 seconds. Otherwise, resume search in 20 seconds. 

The code currently does not have much tests for this features.
A number of functions that are not essential to be viewed in `main.c` have been moved to `callback.c`. `main.c` has been stripped down as far as possible to the core logic, to make reviewing easier.


* put guard

* clean up functions

revert move of print_board_info()

* move function

* move print function

* print board info in seperate file

* move struct declaration to top of file

* use pointer

* move buffer fill to another file

* move board setup to seperate file

* change order

* refactor name of function for readablity

* static internal functions

* move constants to config file

* use more descriptive function name

* move OnRxData() to second file

* remove unused imports

* refactor name of function

* do join as expected

we are in abp mode so it will pass it

* set tx interval depending on fix status

* don't start timers for tx outside the txprepare function

* remove redundant code

* update comment

* remove unused code

* remove redundent variable and declaration

* clean up order

* use #defines to set tx intervals

Update config.h