Don't tx with wrong frequencies (#268)
Add logic to prevent tracker from transmitting if its detected that the LoRaWAN stack has been not initialised to the right regional parameters. If its not been correctly initialized, just don't transmit. It will be initialized correctly later when switching device registrations.

This code adds in a few unittests to test that.