ack successful credential change in EEPROM (#292)
This commit adds a feature to set a bitfield in the telemetry that acks when EEPROM stored credentials have changed from an downlink from the ground. It does significant updates to structures, using fewer points, more atomicity. It adds in a number of unit tests as well, to ensure this change did not break things. The bit fields are combined with the other currently in the telemetry, Playback error. Combining all flags into one set of bit fields simplifies things. The bitfields take the place of the pressure telemetry field, which is no longer used.

This commit also removes unused fields in the `sensor_t` struct that has to do with the GPS. GPS has a separate struct.

undo field removal

its still necessary