Setup gps as soon as the final step in tx is complete (#300)
This commit makes the GPS turn on as soon as a tx/rx is completed. This will ensure that the tracker is drawing come current all the time, somewhat reducing the solar panel voltage. This could protect the electronics somewhat, as the solar panels output too high voltage for the ICs(4.1V max).

The final step is to save to EEPROM the frame count. After which, it will go into deep sleep. Instead of deep sleep, we just turn the GPS back on.

Fixes #283, #265 and #272. By turning on the GPS as soon as a tx/rx is completed, we give the GPS as much time as possible to get a good fix.

Wakeup GPS ONLY after a nvm store call (#309)

Don't wakeup during a nvm restore call.