Randomly pick the network registration to tx on on startup  (#301)
This commit randomly picks the network registration to tx on at startup. This ensures that if it browns out often, it does not rapidly increment the frame count of the first device registration.

The commit sets random seed at boot with the Radio, creates function to set counter randomly and then initializes the  value of the counter soon after boot(before GPS is turned on).

Fix random device credential start (#307)

Set random seed BEFORE using the rand function