change dyanmic model only if it has to be changed (#407)
Change dyanmic model only if it has to be changed. And if it has to be changed, then send all the other GPS setup instructions. When the GPS browns out, it resets all settings, so we need to send them all over again. Earlier, the instructions were sent before each fix attempt, but I suspect this slows down the time to fix. So now, send the instructions only if the settings were incorrect.