Fix a dutycycle-related deadlock in ScheduleTx
When RegionNextChannel returns LORAMAC_STATUS_DUTYCYCLE_RESTRICTED with
a zero DutyCycleWaitTime value, do not return from the function after
skipping the creation of TxDelayedTimer. This would result in a
deadlock. Instead, let the rest of the function to proceed. Only return
if the delayed timer was actually started.

This change ensures that the end of the function with SendFrameOnChannel
has a chance to execute, either immediately when DutyCycleWaitTime == 0,
or later once this function is re-executed when TxDelayedTimer fires.