Only check if TX can be delayed when really needed
When RegionNextChannel returns a zero dutycycle wait time, there is no
need to check if allowDelayedTx is true since the function will attempt
to transmit right away. Thus, the variable only needs to be tested if
there is an actual dutycycle wait time.

If allowDelayedTx is false, abort the function with
LORAMAC_STATUS_DUTYCYCLE_RESTRICTED.