Move LORAMAC[_FALLBACK]?_VERSION macros to the header file (#1291)
* Move LORAMAC[_FALLBACK]?_VERSION macros to the header file

This trivial change makes the two version values accessible to the
application using LoRaMac-node. Being able to obtain the most recent
supported MAC version is useful in case the application needs to
generate data, e.g., for backend (TTN) provisioning.

* Remove ifndef guard surrounding LORAMAC_VERSION

The guard appears to be a leftover from a previous library version and
is no longer needed.