Fix the computation of Block A in FOptsEncrypt
See Section 4.3.1.6 "Frame options (FOptsLen in FCtrl, FOpts)" in
LoRaWAN 1.1 Specification.

The direction parameter is found at index 5 and its value should be 0
for uplinks and 1 for downlinks. The last byte of Block A should be 0
according to Figure 15 in that section.

Tested against The Things Network network server v3.18.2