Fix a dutycycle-related deadlock in ScheduleTx (#1287)
* Fix a dutycycle-related deadlock in ScheduleTx

When RegionNextChannel returns LORAMAC_STATUS_DUTYCYCLE_RESTRICTED with
a zero DutyCycleWaitTime value, do not return from the function after
skipping the creation of TxDelayedTimer. This would result in a
deadlock. Instead, let the rest of the function to proceed. Only return
if the delayed timer was actually started.

This change ensures that the end of the function with SendFrameOnChannel
has a chance to execute, either immediately when DutyCycleWaitTime == 0,
or later once this function is re-executed when TxDelayedTimer fires.

* Only check if TX can be delayed when really needed

When RegionNextChannel returns a zero dutycycle wait time, there is no
need to check if allowDelayedTx is true since the function will attempt
to transmit right away. Thus, the variable only needs to be tested if
there is an actual dutycycle wait time.

If allowDelayedTx is false, abort the function with
LORAMAC_STATUS_DUTYCYCLE_RESTRICTED.