Fixes for class C activation and deactivation (#1325)
* Deactivate radio on a switch from class C to A

When the MAC switches from class C to class A, we must update one of the
internal state variables to make sure that the radio is not put into the
continuous RX mode again in LoRaMacProcess.

* Reactivate radio for continuous RX when starting in class C

* Put radio to sleep in LoRaMacStop if in class C

If the MAC is operating in class C and the application invokes
LoRaMacStop, make sure the radio will not remain receiving continuously,
draining batteries. LoRaMacStart will switch the radio into a continuous
RX mode again.